Split FrontendHelper into several smaller helpers
This way it's going to be easier to manage the code and easier for developers to customize/overwrite those helpers/methods
